[YS-447] refactor: postDetailContent 컴포넌트 로직 유틸로 분리
